incorporate intersection with an interval file
JointHist keys have an additional Option[Boolean] in front that, if
present, indicates a number of loci that either coincided with
(Some(true)) or diverged from (Some(false)) an optionally-provided
interval file. JointHistogram entries with a None in this field can be
taken to represent sums across both, similarly to other fields.